+++ parsley-core/ChangeLog.md
@@ -60,4 +60,8 @@ this meets requirements for internal major change.
 * Refactored `LetBinding` to include more generic metadata.
 * Added metadata to `StaSubroutine` and introduced `StaSubroutine#` and associated functions.
 * Fed metadata through `letRec`'s `genBinding` and into `buildRec`.
-* Added `Known` and `Unknown` `Amount`s to `Offset` to account for the `o+n` case.
+* Added an `Amount` to `Offset`, which also takes into account a multiplicity, used to track unknown
+  but non-zero quantities.
+* Added `callCC` and modified the API for `suspend` to allow for abstracted `Offset` creation. The
+  `callCC` operation promises to utilise static input consumption from the subroutine to refine the
+  input to the return continuation (making use of the multiplicity above).